% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ship_detection.R
\name{cluster_bright_pixels}
\alias{cluster_bright_pixels}
\title{Cluster Bright Pixels Locally}
\usage{
cluster_bright_pixels(bright_pixels, window_size = 51)
}
\arguments{
\item{bright_pixels}{Logical SpatRaster from \code{detect_ships()}.}

\item{window_size}{Integer. Size of the moving window (must be odd).}
}
\value{
Numeric SpatRaster of local bright-pixel counts.
}
\description{
Uses a moving window to count bright pixels in the neighborhood to find clusters.
}
