% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ship_detection.R
\name{filter_clusters}
\alias{filter_clusters}
\title{Filter Pixel Clusters}
\usage{
filter_clusters(local_count, min_cluster_size = 50)
}
\arguments{
\item{local_count}{A numeric SpatRaster of local bright-pixel counts.}

\item{min_cluster_size}{Integer. Minimum cluster size.}
}
\value{
Logical SpatRaster with valid clusters.
}
\description{
Filters out bright pixel clusters smaller than a given size threshold.
}
